+++ 3rdparty/zint-2.6.1/backend/reedsol.c
@@ -120,7 +120,7 @@ void rs_encode(const size_t len,const unsigned char *data, unsigned char *res) {
     int i, k, m;
     for (i = 0; i < rlen; i++)
         res[i] = 0;
-    for (i = 0; i < len; i++) {
+    for (i = 0; i < (int)len; i++) {
         m = res[rlen - 1] ^ data[i];
         for (k = rlen - 1; k > 0; k--) {
             if (m && rspoly[k])
