+++ 3rdparty/zint-2.6.1/backend/svg.c
@@ -83,7 +83,7 @@ int svg_plot(struct zint_symbol *symbol) {
             case BARCODE_UPCE_CC:
                 /* For these symbols use dummy text to ensure formatting is done
                  * properly even if no text is required */
-                for (i = 0; i < ustrlen(symbol->text); i++) {
+                for (i = 0; i < (int)ustrlen(symbol->text); i++) {
                     if (symbol->text[i] == '+') {
                         local_text[i] = '+';
                     } else {
@@ -194,7 +194,7 @@ int svg_plot(struct zint_symbol *symbol) {
     r = 0;
     /* Isolate add-on text */
     if (is_extendable(symbol->symbology)) {
-        for (i = 0; i < ustrlen(local_text); i++) {
+        for (i = 0; i < (int)ustrlen(local_text); i++) {
             if (latch == 1) {
                 addon[r] = local_text[i];
                 r++;
@@ -207,7 +207,7 @@ int svg_plot(struct zint_symbol *symbol) {
     addon[r] = '\0';
 
     /* Don't include control characters in output text */
-    for(i = 0; i < ustrlen(local_text); i++) {
+    for(i = 0; i < (int)ustrlen(local_text); i++) {
         if (local_text[i] < ' ') {
             local_text[i] = ' ';
         }
